change carts migration

+++ db/migrate/20240523091847_remove_discounts_from_carts.rb
@@ -1,5 +1,7 @@
 class RemoveDiscountsFromCarts < ActiveRecord::Migration[7.1]
   def change
-    remove_column :carts, :discount, :decimal
+    if column_exists?(:carts, :discount)
+      remove_column :carts, :discount, :decimal
+    end
   end
 end
